+++ resourcet-extra/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Version 0.0.2
 
-* Add `runResourceT`.
+* Add `runResourceT`, `asyncRestore`.
 
 
 # Version 0.0.1